requirements.txt
# HyperNode-Provisioner Dependencies
# Python dependencies for backend service

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Redis
redis==5.0.1

# Jinja2 for templating
jinja2==3.1.2

# YAML support
pyyaml==6.0.1

# JSON schema validation
jsonschema==4.21.1

# Logging
structlog==24.1.0

# Utility libraries
python-dateutil==2.8.2
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Ansible integration (for future enhancements)
ansible-core==2.16.0
